rewrite /gamble in the class format
